canvis

+++ ex_5/ex_5.ino
@@ -1,8 +1,10 @@
 int drive_gb = 100;
-int drive_mb;
+long drive_mb;     // we changed the type from "int" to "long"
+
 void setup()             // run once, when the sketch starts
 {
-  Serial.begin(9600);     // set up Serial library at 9600 bps
+  Serial.begin(9600);    // set up Serial library at 9600 bps
+
   Serial.print("Your HD is ");
   Serial.print(drive_gb);
   Serial.println(" GB large.");
@@ -14,9 +16,11 @@ void setup()             // run once, when the sketch starts
   Serial.println(" Megabytes!");
 }
 
-void loop()     // we need this to be here even though its empty
+void loop()   // we need this to be here even though its empty
 {
 }
+
+
